clippy: clean kernel crate (zero allows bar too_many_arguments)
~296 of 298 clippy diagnostics fixed across the metal lib, hosted lib, and
hosted bin configs; normal builds green. Validated at runtime: hosted boot
reaches DN (DOS personality + EMS) and a 32-bit Linux ELF runs through the
Linux personality. No logic/numeric/MMIO/layout changes.

Notable non-mechanical fixes (see also the static-mut raw-pointer rule):
- static mut sites (BOOT_MMAP/BOOT_INFO, fbcon GEOM/SHADOW/PALETTE/..., dfs
  CI_CACHE, arch::REGS) bind a &raw mut/const local then deref it.
- #[cfg] gating of config-specific items: platform.rs per-backend enum
  variants (metal VgaCard/Framebuffer/NativeBios/Debugcon, hosted Interp/
  HostWindow/HostStdout), stacktrace::stack_trace, klog metal 0xE9 sink.
- module_inception: inner  -> .
- non_contiguous_range_endpoints in dos.rs done with an inclusive
  ..=SLOT_CB_ENTRY_LAST (the naive 4..=19 would have shadowed INT 13h).
- Default impls forwarding to new() (LinuxState, VgaState); dead code removed
  only where unused in BOTH configs.

Remaining: 2 large_enum_variant (KernelAction::ForkExec, Personality::Dos)
pending a box-vs-allow decision.

+++ kernel/src/kernel/dos/machine/vdma.rs
@@ -98,7 +98,7 @@ impl Dma8237 {
         }
         let hi = port >= 0xC0;
         // Normalize controller #2's ×2 register stride to 0..0x0F.
-        let reg = if hi { ((port - 0xC0) >> 1) as u16 } else { port };
+        let reg = if hi { (port - 0xC0) >> 1 } else { port };
         let chan_base = if hi { 4 } else { 0 };
         match reg {
             0x00..=0x07 => {
@@ -150,7 +150,7 @@ impl Dma8237 {
             return self.ch[chan].prog.page;
         }
         let hi = port >= 0xC0;
-        let reg = if hi { ((port - 0xC0) >> 1) as u16 } else { port };
+        let reg = if hi { (port - 0xC0) >> 1 } else { port };
         let chan_base = if hi { 4 } else { 0 };
         match reg {
             0x00..=0x07 => {
